Increase the size of the socket input buffer beyond a single packet and optimise pending check in TLS.
Pass a hint to count_pending() with the minimum you're looking for. If we already have that in the buffer, answer immediately.
Also make the input buffer a bit more generous. Hopefully these will reduce the number of I/O operations that actually happen.

We could avoid a memmove() if we used a circular input buffer. But let's see if it's a bottleneck first, as it will complicate the code.